+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -75,7 +75,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-26T14:07:37.876Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-27T14:19:20.121Z")
     r"""The date and time the resource was created"""