+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -579,6 +579,9 @@ GetAllPromptsUseCases = Literal[
     "SQL",
     "Summarization",
     "Tagging",
+    "Translation (document)",
+    "Translation (sentences)",
+    "Sentiment analysis",
 ]
 
 GetAllPromptsLanguage = Literal[