+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -559,6 +559,9 @@ GetOnePromptUseCases = Literal[
     "SQL",
     "Summarization",
     "Tagging",
+    "Translation (document)",
+    "Translation (sentences)",
+    "Sentiment analysis",
 ]
 
 GetOnePromptLanguage = Literal[