+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -587,6 +587,9 @@ ListPromptVersionsUseCases = Literal[
     "SQL",
     "Summarization",
     "Tagging",
+    "Translation (document)",
+    "Translation (sentences)",
+    "Sentiment analysis",
 ]
 
 ListPromptVersionsLanguage = Literal[