+++ packages/orq-rc/src/orq_ai_sdk/sdk.py
@@ -13,6 +13,7 @@ from orq_ai_sdk.datasets import Datasets
 from orq_ai_sdk.deployments_sdk import DeploymentsSDK
 from orq_ai_sdk.feedback import Feedback
 from orq_ai_sdk.files import Files
+from orq_ai_sdk.knowledge import Knowledge
 from orq_ai_sdk.models import internal
 from orq_ai_sdk.prompts import Prompts
 from orq_ai_sdk.remoteconfigs import Remoteconfigs
@@ -33,6 +34,7 @@ class Orq(BaseSDK):
     prompts: Prompts
     remoteconfigs: Remoteconfigs
     datasets: Datasets
+    knowledge: Knowledge
 
     def __init__(
         self,
@@ -147,6 +149,7 @@ class Orq(BaseSDK):
         self.prompts = Prompts(self.sdk_configuration)
         self.remoteconfigs = Remoteconfigs(self.sdk_configuration)
         self.datasets = Datasets(self.sdk_configuration)
+        self.knowledge = Knowledge(self.sdk_configuration)
 
     def __enter__(self):
         return self